doc, tools: extract parseLists
